refactor: move seed file to scripts
